+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - New evaluation metrics;
 - New evaluation on test set;
+- New inference service with image upload;
+- Bug fixes;
 
 # Version 0.16.4